Cambio de tipo de dato del stress(int -> double) en NotificacionDetalle.

+++ app/src/main/java/com/example/vitalsignscheckup/NotificacionDetalle.java
@@ -1,18 +1,12 @@
 package com.example.vitalsignscheckup;
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.content.Intent;
 import android.os.Bundle;
 import android.widget.TextView;
 
-import com.example.vitalsignscheckup.models.Mediciones;
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.database.ChildEventListener;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
@@ -20,7 +14,7 @@ public class NotificacionDetalle extends AppCompatActivity {
 
     private double mTemp;
     private int mHeartRate;
-    private int mStress;
+    private double mStress;
     private int mBP1;
     private int mBP2;
 
@@ -31,9 +25,9 @@ public class NotificacionDetalle extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         Intent intent = getIntent();
-        mTemp = intent.getDoubleExtra("Temp", 0);
+        mTemp = intent.getDoubleExtra("Temp", 0.0);
         mHeartRate = intent.getIntExtra("HR", 0);
-        mStress = intent.getIntExtra("Stress", 0);
+        mStress = intent.getDoubleExtra("Stress", 0.0);
         mBP1 = intent.getIntExtra("BP1", 0);
         mBP2 = intent.getIntExtra("BP2", 0);
